refactor(ui): extract lightbox logic to its own component, show close button on small viewports when zooming
